opencvanalyzerbloblocal/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

pandas
azure-functions
opencv-python # Full UI - only for debugging
#opencv-python-headless # For Azure when deployed
ultralytics 
azure-storage-blob